bugfix always do subscription date change on hook
changed subscription dates update to do BOTH a real-time subscription date update and then also schedule it on the hook since we do not know whether the hook will run if the status is or is not being changed. It is VERY IMPORTANT to perform the date change after the hook runs because when woo subscriptions changes the subscription status to Active it will change the dates.